pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = ""
authors = ["zai <artziga77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-star-ratings = "^0.9.2"
geopy = "^2.4.1"
django-formtools = "^2.5.1"
pillow = "^10.1.0"
django-autoslug = "^1.9.9"
django-imagekit = "^5.0.0"
dlib = "^19.24.2"
opencv-python = "^4.8.1.78"
unidecode = "^1.3.7"
django-multiupload = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
